% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annualizer.R
\name{annualizer}
\alias{annualizer}
\title{Annualize results with less than 12 months of observations.}
\usage{
annualizer(x, nmonths)
}
\arguments{
\item{x}{A vector of results to annualize.}

\item{nmonths}{Either a single number that represents the number of months
used for observations, or a vector of length(x) that represents the number
of months each result was observed for.}
}
\value{
vector of length \code{length(x)} with results annualized.
}
\description{
\code{annualizer} performs naive annualization of results with less than
12 months worth of observations, essentialy multiplying each result by
\emph{12/nmonths} for results < 12 months.
}
\examples{
x <- rep(10,20)
 n <- seq(1,20)
 annualize(x, n)
}

