% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findLevel.R
\name{findLevel}
\alias{findLevel}
\title{Find level given minimum threshold}
\usage{
findLevel(x, t)
}
\arguments{
\item{x}{A vector of results to classify.}

\item{t}{A vector of minimum level thresholds, sorted increasingly.}
}
\value{
vector of length \code{length(x)} with values \code{1:(N+1)}.
}
\description{
\code{findLevel} is a wrapper for the \code{findInterval} function
 in base R. Given a vector of results x, and a set of thresholds
 \emph{t1..tn}, it returns a vector of levels with the same length as x,
 with levels \emph{1 to n+1}. The lowest level is 1, and any result acheiving
 >=t1 is given level 2, any result acheiving >=t2 is given level 3 and so on.
}
\examples{
x <- 1:100
 t <- c(10, 20, 50)
 findLevel(x, t)
}

