% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findPayout.R
\name{findPayout}
\alias{findPayout}
\title{Find payout given a result and a payout grid.}
\usage{
findPayout(x, p)
}
\arguments{
\item{x}{A vector of levels to translate to a payout.}

\item{p}{A vector of payout thresholds, sorted increasingly.}
}
\value{
vector of length \code{length(x)}.
}
\description{
\code{findPayout} determines the payout for each level based on a payout grid.
Essentially, for a vector of results against a vector of payouts \emph{p1..pn}
it returns \emph{p[l]} where l is the level, presumably found through
\code{findLevel}. Note that while \code{findLevel} potentially results in one
level greater than the number of thresholds, \code{findPayout} requires a
payout for every level in the vector x.
}
\examples{
x <- c(1,1,2,4)
 p <- c(0, 500, 1000, 2000)
 findPayout(x, p)
 # [1]    0    0  500 2000
\dontrun{
# Not enough payouts for the number of levels provided
x <- sample(rep(1:5, 20))
findPayout(x, p)
}
}

