% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_payout.R
\name{make_payout}
\alias{make_payout}
\title{Create payout grid using target threshold}
\usage{
make_payout(n, target_payout, first_payment = 0)
}
\arguments{
\item{target_payout}{}
}
\description{
Create payout grid using target threshold
}

