% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population.R
\name{pop}
\alias{pop}
\title{Population Object}
\usage{
pop(id, result, group = NULL, scale = 1, tenure = NULL,
  annualize = FALSE, threshold = NULL, paygrid = NULL)
}
\arguments{
\item{annualize}{}
}
\description{
Population Object
}

