% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmdl.R
\name{rmdl.default}
\alias{rmdl.default}
\title{Reward Modelling}
\usage{
\method{rmdl}{default}(id, group = NA, result, threshold = NA,
  payout_grid = NA)
}
\arguments{
\item{id}{unique identifier representing individuals being compensated}

\item{group}{optional grouping for individual employees, e.g. region, market}

\item{result}{the result of the compensable event}

\item{threshold}{a numeric sorted increasingly against which results are compared}

\item{payout_grid}{a numeric sorted of length: length(threshold) + 1 against which results are paid}
}
\value{
An object of class \code{"rmdl"}, which is a list with the following components:
  id, group, result, level, payout
}
\description{
Reward Modelling
}

